% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barnard.R
\name{Barnard}
\alias{Barnard}
\title{Barnard for proportional testing}
\usage{
Barnard(data, Tbx = 100, to.plot = TRUE, to.print = TRUE)
}
\arguments{
\item{data}{A 2x2 matrix}

\item{Tbx}{Number of matricies to test}

\item{to.plot}{TRUE or FALSE to plot}

\item{to.print}{TRUE or FALSe to print the text output}
}
\examples{
Barnard(data=dm, Tbx=200, to.plot=TRUE, to.print=TRUE)
}

