% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2plot.R
\name{deseq2plot}
\alias{deseq2plot}
\title{Make a plot from deseq results}
\usage{
deseq2plot(sigtab, tax)
}
\arguments{
\item{sigtab}{A dataframe with results from DESeq}

\item{tax}{A string with the taxon level to show}
}
\value{
A plot of the sigtab data
}
\description{
Make a plot from deseq results
}
\examples{
deseq2plot(sigtab=sigtab1,tax="Genus")
}

