% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIntersectinos.R
\name{findIntersections}
\alias{findIntersections}
\title{Performs intersection and difference calculations for a list of vectors. 
Performs Fisher exact tests.}
\usage{
findIntersections(x)
}
\arguments{
\item{x}{A list of vectors to compare.}
}
\description{
Performs intersection and difference calculations for a list of vectors. 
Performs Fisher exact tests.
}
\examples{
findIntersections(x)  
}

