% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersMRobject.R
\name{fishersMRobject}
\alias{fishersMRobject}
\title{Performs a Fisher's exact test for 2 outcomes from a Metagenomeseq object}
\usage{
fishersMRobject(obj, otulist, classIndex, norm = TRUE, thresh = 0)
}
\arguments{
\item{obj}{A metagenomeseq object}

\item{otulist}{A list of the otu that should be used in the test.}

\item{classIndex}{A list of samples in each of the test groups}

\item{norm}{TRUE or FALSE to use normalized data.}
}
\value{
Returns a p value for a Fisher exact test.
}
\details{
Sometimes differences between positive and negative measures in samples between groups require statistical evaluation. This can be done using the Chi Square or Fisher's exact test.
}
\examples{
otulist=which(fData(obj)[rank]==testtax)
classIndex=list(unexposed=which(pData(objall.controls)$hiv_exp=="HIVUnexposed"))
classIndex$exposed=which(pData(objall.controls)$hiv_exp=="HIVExposed")
fishersMRobject<-function (obj=objall.controls, otulist=otulist,classIndex=classIndex, norm = TRUE)
}

