% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSSAnova.R
\name{fitSSAnova}
\alias{fitSSAnova}
\title{Performs smoothing splines anova fitting}
\usage{
fitSSAnova(formula, abundance, class, time, id, include = c("class",
  "time:class"))
}
\arguments{
\item{formula}{Formula to use in the SS ANOVA}

\item{abundance}{Abundance data vector}

\item{class}{Class membership (as a factor)}

\item{time}{Time points vector of relative times (same length as abundance vector)}

\item{id}{Sample or Subject ID}

\item{include}{Parameters to include in the SS ANOVA model}
}
\value{
A list
}
\details{
Smoothing-splines ANOVA fit.
}

