% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTimeSeries.2.R
\name{fitTimeSeries.2}
\alias{fitTimeSeries.2}
\title{Performs smoothing splines anova fitting for time series data.}
\usage{
fitTimeSeries.2(melted, abundance = "Abundance", lvl = "moduleColor",
  feature, class, time = "Time", id,
  formula = formula("abundance ~ time * class"), include = c("class",
  "time:class"), C = 0, B = 1000)
}
\arguments{
\item{abundance}{Abundance column name}

\item{lvl}{Level from which to pull features}

\item{feature}{Feature of interest within the lvl}

\item{class}{Class membership (as a factor)}

\item{time}{Time points vector of relative times (same length as abundance vector)}

\item{id}{Sample or Subject ID}

\item{formula}{Formula to use in the SS ANOVA}

\item{include}{Parameters to include in the SS ANOVA model}
}
\value{
A list
}
\details{
Smoothing-splines ANOVA fit using a melted dataframe with abundances and times.
}
\examples{

data(mouseData)
om<- objmelt(mouseData)
om.ddp<- ddply(om, c("Time", "Group","Genus"), summarise, Abundance=sum(Abundance))
fitTimeSeries.2(om.ddp, lvl="Genus",feature="Lactobacillus", class="Group")

}

