% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prop.R
\name{fnct}
\alias{fnct}
\title{plotProp function}
\usage{
fnct(x)
}
\arguments{
\item{x}{The X variable to use}

\item{phy}{a Phyloseq S4 object}

\item{y}{The y variable to use}

\item{fill}{The category to use for making the proportions}

\item{n}{The number of top taxa in the fill category to return}
}
\description{
This function takes input of a phyloseq object, the x and y groups, and the group to use for the fill and returns a  proportion table
}
\examples{
pp<-plotProp(phy1,x="Subject", y="Abundance", fill="Genus", n=10)
}

