% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate the geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Set to default TRUE}
}
\value{
The geometric mean of the vector
}
\description{
Calculate the geometric mean
}
\examples{
gm_mean(list.x,na.rm=TRUE)
}

