% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_richness_map.R
\name{make.richness.map}
\alias{make.richness.map}
\title{Make a map of diversity scores for plotting}
\usage{
make.richness.map(physeq, x = "samples", color = NULL, shape = NULL,
  title = NULL, scales = "free_y", nrow = 1, shsi = NULL,
  measures = NULL, sortby = NULL)
}
\arguments{
\item{physeq}{An S4 phyloseq object}

\item{x}{A string used to designate the x axis for plotting}

\item{color}{A string for the group used to designate colors used in the plot}

\item{shape}{A string for the group used to designate shapes used in the plot}

\item{title}{A string used for the plot title}

\item{scales}{A string for how to scale the axes}

\item{nrow}{A number for how many rows to plot}

\item{shsi}{}

\item{measures}{A string of the diversity indicies to use}

\item{sortby}{}
}
\value{
A dataset for use in a ggplot graph
}
\description{
Make a map of diversity scores for plotting
}
\examples{
make.richness.map(phy,x="samples",color=NULL,shape=NULL,title="A graph",scales= "free_y", nrow = 1, shsi = NULL, measures = NULL, sortby = NULL)  
}

