% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biom_tax_table.R
\name{make_biom_tax_table}
\alias{make_biom_tax_table}
\title{make_biom_tax_table}
\usage{
make_biom_tax_table(phy)
}
\arguments{
\item{phy}{a phyloseq object}
}
\description{
To take a phyloseq taxa table and convert to a biom compatible table
}
\examples{
new_biom_table<-make_biom_table(phy1)
}

