% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deseq.R
\name{make_deseq}
\alias{make_deseq}
\title{Make a DESeq S4 object using a user design and a phyloseq S4 object
 
 @param phy A phyloseq S4 object
 @param design A string
 @param design2 A string
 @return A DESeq S4 object
 @export
 @examples
 make.deseq(phy0,"Treatment+Time+Treatment:Time","Time")}
\usage{
make_deseq(phy, design, design2 = "none")
}
\description{
Make a DESeq S4 object using a user design and a phyloseq S4 object
 
 @param phy A phyloseq S4 object
 @param design A string
 @param design2 A string
 @return A DESeq S4 object
 @export
 @examples
 make.deseq(phy0,"Treatment+Time+Treatment:Time","Time")
}

