% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metagenomeSeq.R
\name{make_metagenomeSeq}
\alias{make_metagenomeSeq}
\title{A script to convert a phyloseq S4 object to a 
metagenomeSeq S4 object 
(from http://joey711.github.io/waste-not-supplemental/
simulation-differential-abundance/
simulation-differential-abundance-server.html)}
\usage{
make_metagenomeSeq(physeq)
}
\arguments{
\item{physeq}{An S4 phyloseq object}
}
\value{
A metagenomeSeq object
}
\description{
A script to convert a phyloseq S4 object to a 
metagenomeSeq S4 object 
(from http://joey711.github.io/waste-not-supplemental/
simulation-differential-abundance/
simulation-differential-abundance-server.html)
}
\examples{
obj<-make_metagenomeSeq(obj)
}

