% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phyloseq.R
\name{make_phyloseq}
\alias{make_phyloseq}
\title{Make phyloseq S4 object}
\usage{
make_phyloseq(workdir, biom, tre, rep.set, map)
}
\arguments{
\item{workdir}{Location of files}

\item{biom}{The biom file name}

\item{tre}{The tree file (midpoint rooting)}

\item{rep.set}{The representative sequence file (rep_set1.fna)}

\item{map}{The metadata map (change '#SampleID' to 'SampleID')}
}
\value{
An S4 phyloseq object
}
\description{
Make phyloseq S4 object
}
\examples{
make_phyloseq(setdir,biom.file,rep_set_tree.file,rep_set.fna,map_file)
}

