% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{make_plsda}
\alias{make_plsda}
\title{This is a function to make the PLDA matrix using 
the mixOmics package.}
\usage{
make_plsda(phy, category, sparse = FALSE)
}
\arguments{
\item{phy}{A phyloseq S4 object}

\item{category}{A grouping category to contrast}
}
\value{
A list and dataframe with the PLSDA information
}
\description{
This is a function to make the PLDA matrix using 
the mixOmics package.
}
\examples{
plsda.phy<-make_plsda(phy0,"Group")
}

