% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxa_tables.R
\name{make_taxa_tables}
\alias{make_taxa_tables}
\title{A function to organize a phyloseq S4 object into 
tables for use in classification modelling}
\usage{
make_taxa_tables(phy)
}
\arguments{
\item{phy}{A phyloseq S4 object}
}
\value{
A dataTable list with per taxa category dataframes of counts
}
\description{
A function to organize a phyloseq S4 object into 
tables for use in classification modelling
}

