% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR2phy.R
\name{MR2phy}
\alias{MR2phy}
\title{MR2phy}
\usage{
MR2phy(obj, norm = TRUE, log = FALSE, taxa_are_rows = TRUE)
}
\arguments{
\item{obj}{A metagenomeseq object}

\item{norm}{TRUE/FALSE if normalized data should be put into phy object}

\item{log}{TRUE/FALSE if log transformated data should be put into phy object}
}
\value{
A phy object
}
\description{
A script to convert a metagenomeseq (Biobase) object to a phyloseq object.
}
\examples{
newphy<-MR2phy(obj)
}

