% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objmelt.R
\name{objmelt}
\alias{objmelt}
\title{Melts an MRexperiments object into a single dataframe.}
\usage{
objmelt(obj.in, norm = TRUE, log = FALSE, sl = 1000)
}
\arguments{
\item{norm}{TRUE-FALSE to use normalized data}

\item{log}{TRUE-FALSE to use log transformed data}

\item{sl}{An integer to use to expand the counts}

\item{obj}{A MRexperiment object.}
}
\value{
A long data.frame
}
\details{
Using the an MRexperiment, the phenotype and count data will be combined into a single long dataframe object.
}
\examples{

data(mouseData)
om <- objmelt(mouseData, norm=TRUE, log=FALSE, sl=1000)

}

