% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu2melt.R
\name{otu2melt}
\alias{otu2melt}
\title{A function to melt data from an S4 phyloseq table with metadata}
\usage{
otu2melt(phy.in)
}
\arguments{
\item{phy.in}{An S4 phyloseq object}
}
\value{
A dataframe with the merged data
}
\description{
A function to melt data from an S4 phyloseq table with metadata
}
\examples{
otu2melt(phy0)
}

