% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy2MR.R
\name{phy2MR}
\alias{phy2MR}
\title{Make a metagenomeSeq object from a phyloseq object}
\usage{
phy2MR(physeq, p = 0.5)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{p}{A quantile value.}
}
\value{
A metagenomeSeq object.
}
\details{
Takes a phyloseq object and a p value (provided or as cumNormStat) and outputs a metagenomeSeq object.
}
\examples{

obj = phy2MR(phy1, p=0.5)
}

