% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_directory.R
\name{pick_directory}
\alias{pick_directory}
\title{A function to choose the working directory}
\usage{
pick_directory()
}
\value{
A string with the working directory
}
\description{
A function to choose the working directory
}
\examples{
pick_directory()
}

