% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiOTU.R
\name{plotMultiOTU}
\alias{plotMultiOTU}
\title{Plots multiple OTU for different subject groups.}
\usage{
plotMultiOTU(obj, otulist, classIndex, log = TRUE, norm = TRUE,
  jitter.factor = 1, pch = 21, labs = TRUE, xlab = NULL, ylab = NULL,
  jitter = TRUE, ret = FALSE, ...)
}
\arguments{
\item{obj}{A metagenomeseq object}

\item{otulist}{A list of the otu that should be used in the test.}

\item{classIndex}{A list of samples in each of the test groups}

\item{log}{TRUE or FALSE to use log2 transformed data.}

\item{norm}{TRUE or FALSE to use normalized data.}

\item{jitter.factor}{The distance factor to jitter plotted points.}

\item{pch}{Character size for labels}

\item{labs}{TRUE or FALSE to plot labels}

\item{xlab}{A character string for the X axis title}

\item{ylab}{A character string for the y axis title.}

\item{jitter}{TRUE or FALSE to use jitter.factor in plotting points.}

\item{ret}{TRUE or FALSE to return the plotting data.}
}
\value{
Plots OTUs in groups. If ret=TRUE, plotting data are returned.
}
\details{
OTU counts from within a taxon may be plot for different subject groups.
}
\examples{
testax=="__Streptococcus"
otulist=which(fData(obj)[rank]==testtax)
classIndex=list(unexposed=which(pData(objall.controls)$hiv_exp=="HIVUnexposed"))
classIndex$exposed=which(pData(objall.controls)$hiv_exp=="HIVExposed")
plotMultiOTU(objall.controls, otulist=otulist, classIndex=classIndex)
}

