% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_OTU.R
\name{replace_OTU}
\alias{replace_OTU}
\title{Replaces the otu_table in a phyloseq object with a new matrix}
\usage{
replace_OTU(physeq, mat)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{mat}{A matrix of new OTU counts with samples in columns and OTU in rows (and OTU names in rownames).}
}
\value{
A revised phyloseq S4 object with the new OTU counts.
}
\details{
Providing a new matrix and a phyloseq object phy to the function, the old OTU table will be replaced with a new matrix of OTU counts. The remaining parts of the S4 objedct will be adjusted accordingly.
}
\examples{

#' replace_OTU(phy=phylo1, mat=newOTUs) 
}

