% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result2taxa.R
\name{result2taxa}
\alias{result2taxa}
\title{Add taxa from a phyloseq object to the results analysis of a DESeq object}
\usage{
result2taxa(phy, res, alpha)
}
\arguments{
\item{phy}{A phyloseq S4 object}

\item{res}{A DESeq result dataframe}

\item{alpha}{A number for the cutoff of the adjusted p values}
}
\value{
A dataframe with the combined results and taxa
}
\description{
Add taxa from a phyloseq object to the results analysis of a DESeq object
}
\examples{
result2taxa(phy=phy0,res=results,alpha=0.1)
}

