% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_plot.R
\name{richness_plot}
\alias{richness_plot}
\title{Plot diversity data from the make_richness function}
\usage{
richness_plot(MR, category = "Group", color = "light green")
}
\arguments{
\item{MR}{A dataframe with the richness data from Make_richness}

\item{category}{A character string in quotes for what category to use for grouping.}

\item{color}{A string with information about the boxplot color}
}
\value{
A boxplot of the richness data
}
\description{
Plot diversity data from the make_richness function
}
\examples{
richness_plot(MR, color="light green")

}

