% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdf.R
\name{splitdf}
\alias{splitdf}
\title{A function to split a dataframe into parts by random sampling
to use in machine learning or Random Forest analysis etc.
From: http://www.gettinggeneticsdone.com/2011/02/split-data-frame-into-testing-and.html}
\usage{
splitdf(dataframe, seed = NULL)
}
\arguments{
\item{dataframe}{The dataframe to split}

\item{seed}{A numeric value for seed reproducibility}
}
\value{
A list containing object$trainset and object$testset
}
\description{
A function to split a dataframe into parts by random sampling
to use in machine learning or Random Forest analysis etc.
From: http://www.gettinggeneticsdone.com/2011/02/split-data-frame-into-testing-and.html
}

