% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMR.R
\name{subset_MR}
\alias{subset_MR}
\title{Subsets a metagenome object by a sample variable like subset does}
\usage{
subset_MR(obj, feature)
}
\arguments{
\item{obj}{A metagenomeseq  objects (from Biobase)}

\item{feature}{Conditions for subsetting in the form .(xxx)}
}
\examples{
newObj<-subset_MR(obj,sampleType=="red")
}

