% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_metagenomeSeq.R
\name{test_metagenomeSeq}
\alias{test_metagenomeSeq}
\title{A function to perform differential abundance analysis on a metagenomeSeq
object (maybe generated from a phyloseq object using the make_metagenomeSeq() script)}
\usage{
test_metagenomeSeq(MGS, variable, physeq = NULL)
}
\arguments{
\item{MGS}{A metagenomeSeq object}

\item{variable}{A phenoData or phyloseq sample_data column variable}

\item{physeq}{A phyloseq object}
}
\value{
A table of p values and p adjusted values for differential abundance
}
\description{
A function to perform differential abundance analysis on a metagenomeSeq
object (maybe generated from a phyloseq object using the make_metagenomeSeq() script)
}
\examples{
t<-test_metagenomeSeq(obj,"Group",phy)
}

