% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimNplot.R
\name{trimNplot}
\alias{trimNplot}
\title{This function will take an S4 metagenomeSeq object and make a proportional barplot
from it.}
\usage{
trimNplot(obj, ftc = "Group", category = "Genus", cutoff = 0,
  out = "matrix", norm = TRUE, log = FALSE)
}
\arguments{
\item{obj}{An metagenomeSeq object}

\item{ftc}{The category for grouping}

\item{category}{A string with the phylogenetic level to test}

\item{cutoff}{A number for the proportion cutoff in the plot}

\item{out}{The object output (matrix vs. MRExperiement object)}

\item{norm}{TRUE/FALSE for using normalized data, default=TRUE}

\item{log}{TRUE/FALSE to use log scale, default=FALSE}
}
\description{
This function will take an S4 metagenomeSeq object and make a proportional barplot
from it.
}

