% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_for_metagenassist.R
\name{write_for_metagenassist}
\alias{write_for_metagenassist}
\title{Output a phyloseq dataset for analysis in Metagenassist}
\usage{
write_for_metagenassist(phy, base.name)
}
\arguments{
\item{phy}{A phyloseq S4 object}

\item{base.name}{The base name for writing files}
}
\description{
Output a phyloseq dataset for analysis in Metagenassist
}

